requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.2
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
distlib==0.3.1
Django==3.2
django-environ==0.4.5
django-filter==2.4.0
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.4
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql==0.0.2
mysqlclient==2.0.3
Pillow==8.2.0
pycparser==2.20
Pygments==2.8.1
pylint==2.6.0
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
serializers==0.2.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.2
wrapt==1.12.1
